+++ src/Molecules/List/List.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import React from 'react';
-import Flexbox from '../../Atoms/Flexbox/index';
-import Typography from '../../Atoms/Typography/index';
+import { Flexbox } from '../..';
+import { Typography } from '../..';
 import { Props } from './List.types';
 
 const StyledList = styled.ul<Props>`
